[Init] Initisalise thte repositiory.
This project aims at providing all the tools necessary to start
a pose estimator for AR tags using a moncular camera.
It is exlusively based on ROS packages and the Alvar software.

The data and markers folders are intended to allow the project
to work out of the shelf and are not expected to contain
configurations/markers for all the cases.

The project is distributed under 3-clause BSD licence by the
contributors of the "Association des Sciences et Techniques
de l'Université Paul Sabatier"